correction envoyées

+++ passenger/passenger.h
@@ -5,7 +5,7 @@ typedef struct s_passenger{
     char surname[20], name[20], destination[20];
     int age, priority, birthday[3], visa;
     int passport_number, place, nb_bags, isRegistered;
-    unsigned long ticket;
+    unsigned int ticket;
     bag bag[2];
     handbag handbag;
 }passenger;
